backend/requirements.txt
anyio==3.6.2
bcrypt==4.0.1
click==8.1.3
fastapi==0.89.1
h11==0.14.0
idna==3.4
passlib==1.7.4
pydantic==1.10.4
PyJWT==2.6.0
python-dotenv==0.21.1
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.1
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
